fix: setting 페이지 서버/클라이언트 분리
